Refactor class location
